Sécurisation, architecture Vite et nettoyage complet
- Suppression des backdoors PHP (install.php, unzip.php) et assets compilés du dépôt
- Migration Laravel Mix → Vite (standard Laravel 12), build vérifié
- Architecture SPA : toutes les routes web renvoient app.blade.php avec @vite
- Router Vue complété (contact intégré, redirect / → /app)
- Middleware SecurityHeaders : X-Frame-Options, HSTS, nosniff, Referrer-Policy…
- CORS piloté par CORS_ALLOWED_ORIGINS dans .env
- UserSeeder : plus d'email/password hardcodé, lecture via ADMIN_EMAIL/ADMIN_PASSWORD
- .env.example : APP_DEBUG=false, SESSION_SECURE_COOKIE=true, secrets vides
- ContactController : validation renforcée (telephone, dns email, max lengths), réponse JSON pure
- Modèle Contact + migration : ajout champ telephone nullable
- Suppression console.log dans tous les composants Vue
- .gitignore : exclusion build/, public/js/, public/css/, backdoors, debugbar
- docker-compose.yml : variables d'env depuis .env, commentaire local-only
- Suppression des vues Blade inutiles (welcome, contact, projets)

+++ database/migrations/2025_04_13_120136_create_projets_table.php
@@ -25,7 +25,7 @@ class CreateProjetsTable extends Migration
             $table->text('bilan');
             $table->string('type');
             $table->integer('ordre');
-            $table->integer('id_vignette')->index();
+            $table->unsignedBigInteger('id_vignette')->index();
             $table->foreign('id_vignette')->references('id')->on('vignettes')->onDelete('cascade');
             $table->timestamps();
         });
